refactor: Extract GroupEmpty into its own file
